Places increase and decrease score in it's own file
